pyproject.toml
[project]
name = "langchain-space"
version = "0.1.0"
description = "A basics exploration on LangChain."
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain-community>=0.3.16",
    "langchain-openai>=0.3.3",
    "pydantic>=2.10.6",
    "pypdf>=5.2.0",
    "python-dotenv>=1.0.1",
]
